Update Offline Plugin docs
