+++ docusaurus/docs/Android/01-basics/04-getting-started.mdx
@@ -175,7 +175,15 @@ A great chat solution should be capable of caching data and handling network con
 `OfflinePlugin` works best when used together with `StatePlugin`.
 :::
 
-You can add it in the following manner:
+Firstly, you need to add an `offline` dependency:
+
+```groovy {2}
+dependencies {
+    implementation "io.getstream:stream-chat-android-offline:$stream_version"
+}
+```
+
+Then, you can add the plugin in the following manner:
 
 <Tabs>
 <TabItem value="kotlin" label="Kotlin">
