+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/client/docusaurus/Attachments.kt
@@ -30,7 +30,7 @@ class Attachments {
                         // Handle success
                     }
                     is Result.Failure -> {
-                        // Handler error
+                        // Handle error
                     }
                 }
             }
