+++ stream-chat-android-docs/src/main/kotlin/io/getstream/chat/docs/kotlin/client/docusaurus/Optimizations.kt
@@ -20,7 +20,7 @@ class Optimizations {
                     // Handle success
                 }
                 is Result.Failure -> {
-                    // Handler error
+                    // Handle error
                 }
             }
         }
